register: port to CommandManager
